# src/services/users/docs/openapi.yaml
openapi: 3.0.1
info:
  title: User Management API
  version: 1.0.0
servers:
  - url: http://localhost:5000/
paths:
  /signup:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        201:
          description: User created successfully
        400:
          description: Username is already taken
  /login:
    post:
      summary: Log in to the application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: Log-in successful
        401:
          description: Invalid credentials
  /users/{id}:
    put:
      summary: Update a user's details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                surname:
                  type: string
                birthdate:
                  type: string
                  format: date
                email:
                  type: string
                phone:
                  type: string
      responses:
        200:
          description: User updated successfully
        401:
          description: Unauthorized request
        404:
          description: User not found
  /posts:
    post:
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        201:
          description: Post created successfully
        401:
          description: Unauthorized request
    get:
      summary: List posts with pagination
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of posts per page
      responses:
        200:
          description: Posts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        title:
                          type: string
                        content:
                          type: string
                        user_id:
                          type: integer
                  total_count:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        401:
          description: Unauthorized request
  /posts/{id}:
    put:
      summary: Update a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        200:
          description: Post updated successfully
        401:
          description: Unauthorized request
        404:
          description: Post not found
    delete:
      summary: Delete a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Post deleted successfully
        401:
          description: Unauthorized request
        404:
          description: Post not found
    get:
      summary: Get a post by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Post retrieved successfully
        401:
          description: Unauthorized request
        404:
          description: Post not found
  /posts/<post_id>/views:
    post:
      summary: Add a view to a post
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: View added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        401:
          description: Unauthorized request
        404:
          description: Post not found
  /posts/<post_id>/likes:
    post:
      summary: Add a like to a post
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Like added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        401:
          description: Unauthorized request
        404:
          description: Post not found